Fixes Issue 259
Refactored Override to Static within the context of a Port. Term now
reflects closer to what is actually happening. Unless an Input Port
explicitily tells OLA that it supports inherit the option will no
show on the web ui.

Had to update some tests to reflect this. Someone else should really
test this code as well.