configure: add option for user to override checks for Python modules
The check for google.protobuf is inherently flawed for cross-compilation,
as it uses the host Python to check for target modules.

In this case, it is not possible to check for a Pyhon modules altogether,
and we need to rely on the user to properly tell us whether we do or do
not have google.protobuf.

This is done by adding am autoconf cache variable:
    ac_cv_have_pymod_google_protobuf

which can be set to:
  - yes     : force the test to be true
  - no      : force the test to be false
  - (empty) : let the test decide

This is then used in AX_PYTHON_MODULE to override the test.

Currently, only the google.protobuf Python module is checked for, but
the same could be done later on of other Python modules that may be
added in the future, of course.

Fixes configure-errors like:
    http://autobuild.buildroot.net/results/e63/e634d5f0cee3967caffaa6e7ada8c0968e3eff63/build-end.log

Also fixes some under-quotation in the python module check.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>