Change names for new plugin
Rename files in new plugin and change all the names and include
paths to reflect this. Add some basic build system support to only
try and build this if the Linux-specific header I am going to
use is present on the system (should avoid a build failure on
BSD systems).