Python client: remove uses of dict.iteritems()
In Python 2, dict.items() returns a list and dict.iteritems() returns an
iterator on the items. In Python 3, only dict.items() exists and it
returns an iterator.

I ran 2to3 with the 'dict' fix, which does the necessary conversion.
While the code is still compatible with Python 2, it won't be completely
as efficient, since we will now create lists where we used iterators
previously. However, since OLA does not deal with huge lists, I think
the performance hit won't be noticeable.

There are some places where we used to use lists, where we can safely
use iterators. These are addressed in the following patch.