Python client: Add __repr__ to OlaClient objects
This is useful for logging and debugging.

Defining __repr__ instead of __str__ allows to see something useful when
printing a list of objects. With __str__, we see:

  [<ola.OlaClient.Device object at 0x7ff3e20fbad0>]

With __repr__:

  [Device(id="2-1", alias=1, name="ArtNet [127.0.0.1]", plugin_id=2, 4 inputs, 4 outputs)]

By default, __str__ calls __repr__, by it could be overriden if we want
a more user-friendly display.

RDMResponse and RDMNack are modified to match the format.
Here is a preview of what the different versions of RDMResponse
look like:

 * ACK, ACK_TIMER and NACK
    RDMResponse(type=ACK, command_class=GET)
    RDMResponse(type=ACK_TIMER, ack_timer=8400 ms, command_class=GET)
    RDMResponse(type=NACK, reason="Write protect")
 * Response with code != OK
    RDMResponse(error="Transaction number mismatch")