olad: Enhance UniverseInfo protobuf to include patched ports
The protobuf currently only contains the number of input/output ports,
which is not very useful in itself. The only way to get a list of
patched ports for a universe is to iterate on all ports on all devices
and check if the port's universe matches. Having the ports details
directly in the universe could be more useful to the client.

The implementation of OladHTTPServer::JsonUniverseInfo could probably be
simplified this way.