Add a setup.py, use it to install Python libs
This is my first attempt at adding a setup.py to the Python code to make
it a proper package. My initial motivation was that I wanted to get the
Python libs to install to a virtualenv. With the current system, Python
files are installed to ${prefix}/lib/pythonX.Y/site-packages, even when
a Python from a virtualenv is active. When installing using setup.py,
the package is installed in the virtualenv.

Another advantage is the finer-grained control of dependencies as well as
automatic installation of them from PyPI.