Move the JaRule & USB core classes out of the plugins directory.
I've created libs/ for things that don't belong common but should be outside
the plugins. We should move plugins/e131/e131 into here as well.

I've also switched to using a USBDeviceID as the map key, rather than a
pointer. This allows us to reduce from 4 to 1 maps in the AsyncPluginImpl.

Once this is in, I can cleanup & simplify the RemoveWidget path, but this PR
was getting too long as it is...