Stop messing with LDFLAGS, use AM_LDFLAGS instead
Automake defines LDFLAGS (and CFLAGS, and a number of other variables)
as user variables (i.e., reserved for the user, not for the build
system), and provides AM_* alternatives for the build system to mangle
instead.

The advantage of this is that the user can mangle those variables by
running "make LDFLAGS=-lfoo" without having to fiddle with configure,
and without losing whatever flags the configure run added.

In addition, since doing "LDFLAGS=-lfoo ./configure" is also supported,
LDFLAGS is captured at AC_OUTPUT time, and therefore should not be
directly mangled by the configure script; otherwise a user might try to
do something which will break the build.

We do so anyway, which might cause things to break in some corner cases.

Fix this misbehaviour by setting AM_LDFLAGS rather than LDLFLAGS, and
AC_SUBST'ing it into the build system at the end of the configure run.