Add systemd unit file for running olad with systemd in user mode
Since most items are tagged using "uaccess", this is the most
easy way to run the olad daemon, without creating additional
system users.

Note that DynamicUsers can be used in future
along with a SupplementaryGroups=plugdev in order to implement
an olad service that can be spawned off the main systemd process
in charge of the whole system. Since this option is not supported
in 232, that has not been enabled here.