examples/ola_recorder: fix overflow handling and warn on slow system.
Previous commit assumed that unsigned int is less wide than uint64_t when
handling unsigned overflow.

But since InMilliSeconds() returns a signed 64-bit value, discard all the
unsigned math and just do signed delta calculations (even though we are
losing one bit. Shouldn't really matter for the majority of cases,
unless someone's going to play for hundereds of years).

Also show a log message when the delay is negative, implying that the
system might be too slow to keep up with the current file. (We might
want to tune that to be < some negative number instead of just < 0
to avoid spamming too much?).

fixes: b82b97015abd96853cd6122f3f3c9d79bd0ed984