Add a Eurolite serial number check
This MR adds an option 'eurolite_serials' to the usb dmx plugin. This option accepts
a comma separated list of serial numbers that are assumed to be Eurolite mk2 devices
(as long as their product and vendor ids also match).

This allows finding the Eurolite MK2 without enabling the 'enable_eurolite_mk2' option.
The advantage is that is avoids the conflicts that 'enable_eurolite_mk2' produces
with other plugins. These
conflicts make it impossible to connect the mk2 simultaneously together with most other
interfaces to have multiple universes and/or input universes.

I have tested this by connecting an Enttec pro together with the Eurolite Mk2, and this
works.