Add protobuf syntax version
The protocol buffer spec suggests that this is required.
I've tried two protobuf libraries, and one errors without this,
and the other prints a warning.

There's no documentation in OLA I can find that it's proto2,
but it works this way, and it makes sense.  Both files use "required",
which was removed from proto3.